+++ camera.cpp
@@ -5,56 +5,56 @@
 
 namespace mrtp {
 
-Camera::Camera(Eigen::Vector3f *eye, Eigen::Vector3f *lookat, float roll) : eye_(*eye), lookat_(*lookat), roll_(roll) {}
+Camera::Camera(Eigen::Vector3d *eye, Eigen::Vector3d *lookat, double roll) : eye_(*eye), lookat_(*lookat), roll_(roll) {}
 
 Camera::~Camera() {}
 
-void Camera::calculate_window(int width, int height, float perspective) {
+void Camera::calculate_window(int width, int height, double perspective) {
     // i is a vector between the camera and the center of the window
-    Eigen::Vector3f i = lookat_ - eye_;
-    i *= (1.0f / i.norm());
+    Eigen::Vector3d i = lookat_ - eye_;
+    i *= (1 / i.norm());
 
-    Eigen::Vector3f k;
-    k << 0.0f, 0.0f, 1.0f;
+    Eigen::Vector3d k;
+    k << 0, 0, 1;
 
-    Eigen::Vector3f j = i.cross(k);
-    j *= (1.0f / j.norm());
+    Eigen::Vector3d j = i.cross(k);
+    j *= (1 / j.norm());
 
     k = j.cross(i);
-    k *= (1.0f / k.norm());
+    k *= (1 / k.norm());
 
     // Rotate camera around the i axis
-    float roll = roll_ * M_PI / 180.0f;
-    float sina = std::sin(roll);
-    float cosa = std::cos(roll);
+    double roll = roll_ * M_PI / 180;
+    double sina = std::sin(roll);
+    double cosa = std::cos(roll);
 
-    Eigen::Vector3f jp = cosa * j + sina * k;
-    Eigen::Vector3f kp = -sina * j + cosa * k;
+    Eigen::Vector3d jp = cosa * j + sina * k;
+    Eigen::Vector3d kp = -sina * j + cosa * k;
 
     j = jp;
     k = kp;
 
     // Calculate the central point of the window
-    Eigen::Vector3f center = eye_ + perspective * i;
+    Eigen::Vector3d center = eye_ + perspective * i;
 
     // Find three corners of the window
-    wo_ = center - 0.5f * j + 0.5f * k;
+    wo_ = center - 0.5 * j + 0.5 * k;
 
-    Eigen::Vector3f h = wo_ + j;
-    Eigen::Vector3f v = wo_ - k;
+    Eigen::Vector3d h = wo_ + j;
+    Eigen::Vector3d v = wo_ - k;
 
     // Find vectors spanning the window
-    wh_ = 1.0f / static_cast<float>(width) * (h - wo_);
-    wv_ = 1.0f / static_cast<float>(height) * (v - wo_);
+    wh_ = 1 / static_cast<double>(width) * (h - wo_);
+    wv_ = 1 / static_cast<double>(height) * (v - wo_);
 }
 
-Eigen::Vector3f Camera::calculate_origin(int windowx, int windowy) {
-    return (wo_ + static_cast<float>(windowx) * wh_ + static_cast<float>(windowy) * wv_);
+Eigen::Vector3d Camera::calculate_origin(int windowx, int windowy) {
+    return (wo_ + static_cast<double>(windowx) * wh_ + static_cast<double>(windowy) * wv_);
 }
 
-Eigen::Vector3f Camera::calculate_direction(Eigen::Vector3f *origin) {
-    Eigen::Vector3f direction = (*origin) - eye_;
-    return (direction * (1.0f / direction.norm()));
+Eigen::Vector3d Camera::calculate_direction(Eigen::Vector3d *origin) {
+    Eigen::Vector3d direction = (*origin) - eye_;
+    return (direction * (1 / direction.norm()));
 }
 
 } //namespace mrtp
