+++ light.cpp
@@ -3,10 +3,10 @@
 
 namespace mrtp {
 
-Light::Light(Eigen::Vector3f *center) : center_(*center) {}
+Light::Light(Eigen::Vector3d *center) : center_(*center) {}
 
 Light::~Light() {}
 
-Eigen::Vector3f Light::calculate_ray(Eigen::Vector3f *hit) { return (center_ - (*hit)); }
+Eigen::Vector3d Light::calculate_ray(Eigen::Vector3d *hit) { return (center_ - (*hit)); }
 
 } //namespace mrtp
